Add files via upload
This is a general purpose code for handling fasta file. As an example, I have chosen 2019 nCoV - Corona virus genome
Here is the basic stats of #Coronavirus genome (MT020881_1, 29.8kb). The sequence is downloaded from NCBI (Note: It is a ssRNA virus).
I observed that the GC-content of this virus genome is 38% however it is generally high at beginning and the end of the genome. It is possible that the ends of this viral genome (with higher GC content) have more stable secondary structures than central region (with lower GC content).